@@ -13,16 +13,20 @@
     ]
   ],
   "_from": "iconv-lite@>=0.4.13 <0.5.0",
-  "_id": "iconv-lite@0.4.13",
+  "_id": "iconv-lite@0.4.15",
   "_inCache": true,
   "_installable": true,
   "_location": "/iconv-lite",
-  "_nodeVersion": "4.1.1",
+  "_nodeVersion": "7.0.0",
+  "_npmOperationalInternal": {
+    "host": "packages-18-east.internal.npmjs.com",
+    "tmp": "tmp/iconv-lite-0.4.15.tgz_1479754977280_0.752664492232725"
+  },
   "_npmUser": {
     "email": "ashtuchkin@gmail.com",
     "name": "ashtuchkin"
   },
-  "_npmVersion": "2.14.4",
+  "_npmVersion": "2.15.9",
   "_phantomChildren": {},
   "_requested": {
     "name": "iconv-lite",
@@ -35,8 +39,8 @@
   "_requiredBy": [
     "/encoding"
   ],
-  "_resolved": "https://registry.npmjs.org/iconv-lite/-/iconv-lite-0.4.13.tgz",
-  "_shasum": "1f88aba4ab0b1508e8312acc39345f36e992e2f2",
+  "_resolved": "https://registry.npmjs.org/iconv-lite/-/iconv-lite-0.4.15.tgz",
+  "_shasum": "fe265a218ac6a57cfe854927e9d04c19825eddeb",
   "_shrinkwrap": null,
   "_spec": "iconv-lite@~0.4.13",
   "_where": "/tmp/build/88bfaeb5/gpcc/frontend/node_modules/encoding",
@@ -106,21 +110,21 @@
   "devDependencies": {
     "async": "*",
     "errto": "*",
-    "iconv": "2.1",
+    "iconv": "*",
     "istanbul": "*",
     "mocha": "*",
-    "request": "2.47",
+    "request": "*",
     "unorm": "*"
   },
   "directories": {},
   "dist": {
-    "shasum": "1f88aba4ab0b1508e8312acc39345f36e992e2f2",
-    "tarball": "https://registry.npmjs.org/iconv-lite/-/iconv-lite-0.4.13.tgz"
+    "shasum": "fe265a218ac6a57cfe854927e9d04c19825eddeb",
+    "tarball": "https://registry.npmjs.org/iconv-lite/-/iconv-lite-0.4.15.tgz"
   },
   "engines": {
-    "node": ">=0.8.0"
+    "node": ">=0.10.0"
   },
-  "gitHead": "f5ec51b1e7dd1477a3570824960641eebdc5fbc6",
+  "gitHead": "c3bcedcd6a5025c25e39ed1782347acaed1d290f",
   "homepage": "https://github.com/ashtuchkin/iconv-lite",
   "keywords": [
     "iconv",
@@ -148,5 +152,6 @@
     "coverage-open": "open coverage/lcov-report/index.html",
     "test": "mocha --reporter spec --grep ."
   },
-  "version": "0.4.13"
+  "typings": "./lib/index.d.ts",
+  "version": "0.4.15"
 }
