@@ -3,24 +3,24 @@
     [
       {
         "name": "moment",
-        "raw": "moment@2.16.0",
-        "rawSpec": "2.16.0",
+        "raw": "moment@2.17.0",
+        "rawSpec": "2.17.0",
         "scope": null,
-        "spec": "2.16.0",
+        "spec": "2.17.0",
         "type": "version"
       },
       "/tmp/build/88bfaeb5/gpcc/frontend"
     ]
   ],
-  "_from": "moment@2.16.0",
-  "_id": "moment@2.16.0",
+  "_from": "moment@2.17.0",
+  "_id": "moment@2.17.0",
   "_inCache": true,
   "_installable": true,
   "_location": "/moment",
   "_nodeVersion": "6.7.0",
   "_npmOperationalInternal": {
-    "host": "packages-18-east.internal.npmjs.com",
-    "tmp": "tmp/moment-2.16.0.tgz_1478760885663_0.9942573502194136"
+    "host": "packages-12-west.internal.npmjs.com",
+    "tmp": "tmp/moment-2.17.0.tgz_1479822489109_0.9663237682543695"
   },
   "_npmUser": {
     "email": "iskren.chernev@gmail.com",
@@ -30,20 +30,20 @@
   "_phantomChildren": {},
   "_requested": {
     "name": "moment",
-    "raw": "moment@2.16.0",
-    "rawSpec": "2.16.0",
+    "raw": "moment@2.17.0",
+    "rawSpec": "2.17.0",
     "scope": null,
-    "spec": "2.16.0",
+    "spec": "2.17.0",
     "type": "version"
   },
   "_requiredBy": [
     "/",
     "/react-datepicker"
   ],
-  "_resolved": "https://registry.npmjs.org/moment/-/moment-2.16.0.tgz",
-  "_shasum": "f38f2c97c9889b0ee18fc6cc392e1e443ad2da8e",
+  "_resolved": "https://registry.npmjs.org/moment/-/moment-2.17.0.tgz",
+  "_shasum": "a4c292e02aac5ddefb29a6eed24f51938dd3b74f",
   "_shrinkwrap": null,
-  "_spec": "moment@2.16.0",
+  "_spec": "moment@2.17.0",
   "_where": "/tmp/build/88bfaeb5/gpcc/frontend",
   "author": {
     "email": "iskren.chernev@gmail.com",
@@ -111,13 +111,13 @@
     "qunit-cli": "^0.1.4",
     "rollup": "latest",
     "spacejam": "latest",
-    "typescript": "^1.8.10",
+    "typescript": "^2.0.8",
     "uglify-js": "latest"
   },
   "directories": {},
   "dist": {
-    "shasum": "f38f2c97c9889b0ee18fc6cc392e1e443ad2da8e",
-    "tarball": "https://registry.npmjs.org/moment/-/moment-2.16.0.tgz"
+    "shasum": "a4c292e02aac5ddefb29a6eed24f51938dd3b74f",
+    "tarball": "https://registry.npmjs.org/moment/-/moment-2.17.0.tgz"
   },
   "dojoBuild": "package.js",
   "ender": "./ender.js",
@@ -189,5 +189,5 @@
     ]
   },
   "typings": "./moment.d.ts",
-  "version": "2.16.0"
+  "version": "2.17.0"
 }
