@@ -3,24 +3,24 @@
     [
       {
         "name": "react-bootstrap",
-        "raw": "react-bootstrap@0.30.6",
-        "rawSpec": "0.30.6",
+        "raw": "react-bootstrap@0.30.7",
+        "rawSpec": "0.30.7",
         "scope": null,
-        "spec": "0.30.6",
+        "spec": "0.30.7",
         "type": "version"
       },
       "/tmp/build/88bfaeb5/gpcc/frontend"
     ]
   ],
-  "_from": "react-bootstrap@0.30.6",
-  "_id": "react-bootstrap@0.30.6",
+  "_from": "react-bootstrap@0.30.7",
+  "_id": "react-bootstrap@0.30.7",
   "_inCache": true,
   "_installable": true,
   "_location": "/react-bootstrap",
-  "_nodeVersion": "7.0.0",
+  "_nodeVersion": "7.1.0",
   "_npmOperationalInternal": {
     "host": "packages-12-west.internal.npmjs.com",
-    "tmp": "tmp/react-bootstrap-0.30.6.tgz_1477680174956_0.3036094370763749"
+    "tmp": "tmp/react-bootstrap-0.30.7.tgz_1479845049145_0.8373064689803869"
   },
   "_npmUser": {
     "email": "tesrin@gmail.com",
@@ -35,19 +35,19 @@
   },
   "_requested": {
     "name": "react-bootstrap",
-    "raw": "react-bootstrap@0.30.6",
-    "rawSpec": "0.30.6",
+    "raw": "react-bootstrap@0.30.7",
+    "rawSpec": "0.30.7",
     "scope": null,
-    "spec": "0.30.6",
+    "spec": "0.30.7",
     "type": "version"
   },
   "_requiredBy": [
     "/"
   ],
-  "_resolved": "https://registry.npmjs.org/react-bootstrap/-/react-bootstrap-0.30.6.tgz",
-  "_shasum": "288662a245f9dbb79f7740ee595e4ec931d6a4a9",
+  "_resolved": "https://registry.npmjs.org/react-bootstrap/-/react-bootstrap-0.30.7.tgz",
+  "_shasum": "39da80088693ecb71e8e63b5bdc313571fd993d1",
   "_shrinkwrap": null,
-  "_spec": "react-bootstrap@0.30.6",
+  "_spec": "react-bootstrap@0.30.7",
   "_where": "/tmp/build/88bfaeb5/gpcc/frontend",
   "author": {
     "email": "stevoland@gmail.com",
@@ -126,9 +126,10 @@
     "nodemon": "^1.9.2",
     "output-file-sync": "^1.1.2",
     "portfinder": "^1.0.3",
-    "react": "^15.3.0",
-    "react-component-metadata": "^3.0.0",
-    "react-dom": "^15.3.0",
+    "react": "^15.4.0",
+    "react-addons-test-utils": "^15.4.0",
+    "react-component-metadata": "^3.1.0",
+    "react-dom": "^15.4.0",
     "react-hot-loader": "^1.3.0",
     "react-router": "^2.6.1",
     "react-waypoint": "^3.0.0",
@@ -138,7 +139,7 @@
     "sinon": "^1.17.4",
     "sinon-chai": "^2.8.0",
     "style-loader": "^0.13.1",
-    "teaspoon": "^6.4.1",
+    "teaspoon": "^6.4.2",
     "transform-loader": "^0.2.3",
     "webpack": "^1.13.1",
     "webpack-dev-server": "^1.14.1",
@@ -146,8 +147,8 @@
   },
   "directories": {},
   "dist": {
-    "shasum": "288662a245f9dbb79f7740ee595e4ec931d6a4a9",
-    "tarball": "https://registry.npmjs.org/react-bootstrap/-/react-bootstrap-0.30.6.tgz"
+    "shasum": "39da80088693ecb71e8e63b5bdc313571fd993d1",
+    "tarball": "https://registry.npmjs.org/react-bootstrap/-/react-bootstrap-0.30.7.tgz"
   },
   "files": [
     "CHANGELOG.md",
@@ -155,7 +156,7 @@
     "dist",
     "es"
   ],
-  "gitHead": "c84f3031c3b09adce2ef4288ed24235e9c204760",
+  "gitHead": "93814ff8053d4d3c06f8d96edf342430de1e2bf4",
   "homepage": "https://react-bootstrap.github.io/",
   "jsnext:main": "es/index.js",
   "keywords": [
@@ -213,5 +214,5 @@
     "test-browser": "cross-env NODE_ENV=test karma start --single-run",
     "test-node": "mocha --compilers js:babel-register test/server/*Spec.js"
   },
-  "version": "0.30.6"
+  "version": "0.30.7"
 }
