@@ -15,7 +15,7 @@ var propTypes = {
   /**
    * Only valid if `inline` is not set.
    */
-  validationState: React.PropTypes.oneOf(['success', 'warning', 'error']),
+  validationState: React.PropTypes.oneOf(['success', 'warning', 'error', null]),
   /**
    * Attaches a ref to the `<input>` element. Only functions can be used here.
    *
