@@ -45,7 +45,7 @@ var ca = moment.defineLocale('ca', {
         sameElse : 'L'
     },
     relativeTime : {
-        future : 'en %s',
+        future : 'd\'aquí %s',
         past : 'fa %s',
         s : 'uns segons',
         m : 'un minut',
