@@ -10,4 +10,4 @@
 
 'use strict';
 
-module.exports = '15.4.0';
+module.exports = '15.4.1';
